requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic-ai==0.7.4
pandas==2.3.1
openpyxl==3.1.5
PyMuPDF==1.26.3
python-docx==1.2.0
beautifulsoup4==4.13.4
slowapi==0.1.9
tabulate==0.9.0